% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{invlogit}
\alias{invlogit}
\title{inverse logit function}
\usage{
invlogit(x)
}
\arguments{
\item{x}{a real number}
}
\value{
exp(x)/(1+exp(x))
}
\description{
inverse logit function
}
\examples{
invlogit(0.1)
}
