% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{A Model fit Function}
\usage{
model_fit(model = "ITR", tol_iter, thin, mh_sd, K_u = NULL, K_f = NULL, W,
  prior_mu_log_sd2_u0, prior_sigma_log_sd2_u0, prior_sd1_f = 0.01,
  prior_u10_f, prior_sd10_f, prior_sdr_f = 0.09, prior_ur0_f, prior_sdr0_f,
  prior_ub0_f, prior_mu_log_sd2_f0, prior_sigma_log_sd2_f0, prior_unit = 1,
  data, model_fit = NULL, last_iter = 1)
}
\arguments{
\item{model}{Specify which model to fit; models include DPMITR, DPITR and ITR; defaults to ITR}

\item{tol_iter}{Total number of iterations for MCMC sampling}

\item{thin}{Number to thin in MCMC sampling}

\item{mh_sd}{Standard deviation for proposal normal distribution in metropolis-hastings}

\item{K_u}{The max number of clusters for baseline progression; must be specified for model DPMITR and DPITR}

\item{K_f}{1xD, the max number of clusters for each treatment type d; must be specified for model DPMITR and DPITR}

\item{W}{1xD, window size for treatment type d}

\item{prior_mu_log_sd2_u0}{Mean of the sd in U's Gaussian process; defaults to log(1)}

\item{prior_sigma_log_sd2_u0}{Sd of the sd in U's Gaussian process; defaults to 4}

\item{prior_sd1_f}{Sd of the subgroup prior for peak-effect alpha_1; defaults to 0.01}

\item{prior_u10_f}{1xD, mean of the base prior for peak-effect alpha_1, better specified with domain knowledge}

\item{prior_sd10_f}{Sd of the base prior for peak-effect alpha_1; defaults to 0.01}

\item{prior_sdr_f}{Sd of the subgroup priors for time-to-peak-effect gamma; defaults to 0.09}

\item{prior_ur0_f}{1xD, mean of the base prior for time-to-peak-effect gamma, better specified with domain knowledge}

\item{prior_sdr0_f}{Sd of the base prior for time-to-peak-effect gamma; defaults to 4}

\item{prior_ub0_f}{1xD, mean of the base prior for long-term-remaining-effect b, better specified with domain knowledge; if no long term effects, this can be set 0}

\item{prior_mu_log_sd2_f0}{Mean of the sd in the noise Gaussian process}

\item{prior_sigma_log_sd2_f0}{Sd of the sd in the noise Gaussian process; defaults to 4}

\item{prior_unit}{Unit of the prior distributions; defaults to 1}

\item{data}{Input data; please see SimulateData for detailed data format}

\item{model_fit}{Defaults to 'NULL'; otherwise specify the intermediate chain file as the current start}

\item{last_iter}{Defaults to 0; otherwise specify which iterations to continue with when start with an itermediate chain file}
}
\value{
A list of all the parameters defined in the model. \item{sigma2_u}{The individual-level estimated hyperparamters in the Gaussian process for baseline progression.}\item{rho_u}{The individual-level estimated hyperparamters in the Gaussian process for baseline progression.}\item{U}{The individual-level infered mean of the baseline progression.}\item{Beta}{The individual-level estimated Beta coefficients in the liner regression for the basline progression.}\item{a1}{The individual-level estimated peak effect alpha_1 defined in the treatment response curves.}\item{a2}{The individual-level estimated effect increasing slope alpha_2 defined in the treatment response curves.}\item{a3}{The individual-level estimated effect decreasing slope alpha_3 defined in the treatment response curves.}\item{r}{The individual-level estimated time to peak effect gamma defined in the treatment response curves.}\item{b}{The individual-level estimated remaining long-term effect b defined in the treatment response curves.}\item{rho_f}{The individual-level estimated hyperparameter in the Gaussian process prior of the treatment effect noises.}\item{sigma2_f}{The individual-level estimated hyperparameter in the Gaussian process prior of the treatment effect noises.}\item{f}{The indivdiual-level infered mean of the cumulative treatment effects}\item{Sigma_e}{The indivdiual-level estimated mean of the Gaussian noise}\item{ll}{The log-likelihood per iteration.} \item{C_u}{The estimated cluster assignments of baseline progressions; for model DPMITR and DPITR only.}\item{mu_log_sd2_u}{The subgroup-level estimated hyperparameter means in the Gaussian process prior for baseline progression; for model DPMITR and DPITR only.}\item{mu_log_rho_u}{The subgroup-level estimated hyperparameter means in the Gaussian process prior for baseline progression; for model DPMITR and DPITR only.}\item{Mu_b}{The subgroup-level estimated means for the prior of Beta; for model DPMITR and DPITR only.}\item{Sigma_b}{The subgroup-level estimated sd for the prior of Beta; for model DPMITR only.}\item{u1_f}{The subgroup-level estimated mean for the parameters defining treatment response curves; for model DPMITR and DPITR only.}\item{u2_f}{The subgroup-level estimated mean for the parameters defining treatment response curves; for model DPMITR and DPITR only.}\item{u3_f}{The subgroup-level estimated mean for the parameters defining treatment response curves; for model DPMITR and DPITR only.}\item{ur_f}{The subgroup-level estimated mean for the parameters defining treatment response curves; for model DPMITR and DPITR only.}\item{ub_f}{The subgroup-level estimated mean for the parameters defining treatment response curves; for model DPMITR and DPITR only.}\item{C_f}{The estimated cluster assignments for each treatment type d; for model DPMITR and DPITR only.}\item{M1}{The estimated hyperparameter for baseline progression clustering; for model DPMITR and DPITR only.}\item{M2}{The estimated hyperparameter for treatment effect noise clustering; for model DPMITR and DPITR only.}
}
\description{
This function allows you to obtain Bayesian etimation of baseline progressions (i.e. under no treatments) and treatment response curves (with treatments) from longitudinal observational data. Model can be trained only at indivdiual level (ITR), only at subgroup level (DPITR), and at both levels (DPMITR).
}
\examples{
## simulate data
K_u = 3
K_f = c(3,3)
W = c(60,70)

sim_data = SimulateData(I=200, W, trajectory_plot = TRUE)

tol_iter = 100
thin = 10
mh_sd = 0.3
prior_unit = 1

## U kernel
prior_mu_log_sd2_u0 = log(0.01)
prior_sigma_log_sd2_u0 = 0.09
prior_mu_log_sd2_u = c(log(0.01),log(0.01),log(0.01))
prior_mu_logit_rho_u = c(logit(0.1),logit(0.9),logit(0.5))

## f kernel
prior_sd1_f = 0.09
prior_u10_f = c(8,-8)
prior_sd10_f = 4

prior_sdr_f = 0.01
prior_ur0_f = c(10,20)
prior_sdr0_f = 4

prior_ub0_f = c(logit(0.01),logit(0.01))

prior_mu_log_sd2_f0 = log(0.01)
prior_sigma_log_sd2_f0 = 0.09

## run model fit
ptm = proc.time()
chain = model_fit(model = 'DPMITR', tol_iter, thin, mh_sd,
                     K_u, K_f, W,
                     prior_mu_log_sd2_u0, prior_sigma_log_sd2_u0,
                     prior_sd1_f, prior_u10_f, prior_sd10_f,
                     prior_sdr_f, prior_ur0_f, prior_sdr0_f,
                     prior_ub0_f,
                     prior_mu_log_sd2_f0, prior_sigma_log_sd2_f0,
                     prior_unit,
                     data = sim_data,
                     model_fit = NULL,last_iter = 0)
runtime = proc.time() - ptm
chain$runtime = runtime;

## store the final results
save(chain, file = "chain.dpmitr.sim.RData");

}
\references{
Yanbo Xu, Yanxun Xu and Suchi Saria, 2016. A Bayesian Nonparametic Approach for Estimating Individualized Treatment-Response Curves. arXiv preprint arXiv:1608.05182. [tentatively accepted by Journal of Machine Learning Research]
}
\author{
Yanbo Xu
}
\keyword{curves}
\keyword{effects}
\keyword{heterogenous}
\keyword{individualized}
\keyword{response}
\keyword{treatment}
