% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_fit}
\alias{plot_model_fit}
\title{A Plot Function}
\usage{
plot_model_fit(chain, start_iter, tol_iter, prior_unit, train_data, W)
}
\arguments{
\item{chain}{The model fit output from model_fit_dpmitr.}

\item{start_iter}{The thined iteration to start in chain for computing the MCMC estimation.}

\item{tol_iter}{The last thined iteration to end in chain for computing the MCMC estimation.}

\item{prior_unit}{The unit of the prior distribution; defaults to 1.}

\item{train_data}{The training data.}

\item{W}{The treatment effective window}
}
\description{
This function plots the estimated baseline progression and treatment response curves both at the individual level and subgroup level.
}
\examples{
plot_model_fit(chain, start_iter = 5, tol_iter = 9, prior_unit = 1, train_data = sim_data, W)
}
\author{
Yanbo Xu
}
