% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sim_data}
\alias{plot_sim_data}
\title{A Plot Function}
\usage{
plot_sim_data(sim_data)
}
\arguments{
\item{sim_data}{The simulated data}
}
\description{
This function plots the heterogenous baseline progression and treatment response curves generated in the simulated data
}
\examples{
plot_sim_data(sim_data)
}
\author{
Yanbo Xu
}
