% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{A more detailed data simulation function}
\usage{
simulate_data(I, minT, maxT, unit, minDur, maxDur, minADur, maxADur, W,
  prior_unit = 1, trajectory_plot = TRUE)
}
\arguments{
\item{I}{Size of the cohort, i.e. number of trajectories}

\item{minT}{The minimum total duration of the trajectories generated in the data}

\item{maxT}{The maximum total duration of the trajectories generated in the data}

\item{unit}{The time unit of minT and maxT, e.g. unit = 60 means a unit of hour}

\item{minDur}{The minimum duration to generate next observation in a trajectory}

\item{maxDur}{The maximum duration to generate next observation in a trajectory}

\item{minADur}{The minimum duration to generate next treatment management in a trajectory}

\item{maxADur}{The maximum duration to generate next treatment management in a trajectory}

\item{W}{1xD, window size for treatment type d}

\item{prior_unit}{The unit of the prior distributions; defaults to 1}

\item{trajectory_plot}{Defaults to TRUE; TRUE if need to plot upto 20 subsample trajectories from the simulated data.}
}
\value{
A list of the simulated input data and the ground truth parameters that were used for data generation. \item{data_I}{Size of the cohort.}\item{data_D}{Number of treatment types.}\item{data_P}{Number of baseline variables in the linear regression.}\item{data_J_I}{1xI, the number of observations in each trajectory i.}\item{data_J}{The maximum number of observations that can be generated in each trajectory, i.e. max(data_J_I).}\item{data_L_I}{1xI, the number of treatments managed in each trajectory i.}\item{data_L}{The maximum number of treatments that can be generated in each trajectory, i.e. max(data_L_I).}\item{data_T_I}{1xI, the duration of each trajectory i.}\item{data_T}{IxJ, data_T[i,1:J_I[i]] contains the irregular time points at which observations were generated in trajectory i.}\item{data_Tau}{IxL, data_Tau[i,1:L_I[i]] contains the irregular time points at which treatments were managed in trajectory i.}\item{data_B}{IxJxP, the simulated baseline variables in the linear regression.}\item{data_A}{IxL, the sequence of treatment types managed in each trajectory i.}\item{data_Z}{IxJ, the sequene of the final observations for each trajectory i.}\item{data_avg_output}{Defaults to 0; the standardized mean of the data.}\item{data_sd_output}{Defaults to 1; the standardized sd of the data.}\item{param_a1}{IxD, the ground truth parameter for individual-level peak effects alpha_1.}\item{param_a2}{IxD, the ground truth parameter for individual-level effect increasing slope alpha_2.}\item{param_a3}{IxD, the ground truth parameter for individual-level effect decreasing slope alpha_3.}\item{param_r}{IxD, the ground truth parameter for individual-level time to peak effect gamma.}\item{param_b}{IxD, the ground truth parameter for individual-level remaining long-term effect b.}\item{param_C_f}{IxD, the true cluster assignment of the treatment effects in trajectory i.}\item{param_u1_f}{max(K_f)xD, the ground truth subgroup-level mean of the peak effects.}\item{param_u2_f}{max(K_f)xD, the ground truth subgroup-level mean of the effect increasing slope.}\item{param_u3_f}{max(K_f)xD, the ground truth subgroup-level mean of the effect decreasing slope.}\item{param_ur_f}{max(K_f)xD, the ground truth subgroup-level mean of the time to peak effect.}\item{param_ub_f}{max(K_f)xD, the ground truth subgroup-level mean of the remaining long-term effect.}\item{param_Beta}{IxP, the ground truth individual-level linear regression coefficients in the baseline progressions.}\item{param_sigma2_u}{1xI, the ground truth individual-level Gaussian process hyperparameters in the baseline progression}\item{param_rho_u}{1xI, the ground truth individual-level Gaussian process hyperparameters in the baseline progression}\item{param_U}{IxJ, the ground truth individual-level baseline progressions.}\item{param_C_u}{1xI, the true cluster assignment of the baseline progression in trajectory i.}\item{param_Mu_b}{K_uxP, the ground truth subgroup-level means of the linear regression coefficients in the baseline progression}\item{param_mu_log_sd2_u}{1xK_u, the ground truth subgroup-level means of the Gaussian process hyperparameters in the baseline progression}\item{param_mu_logit_rho_u}{1xK_u, the ground truth subgroup-level means of the Gaussian process hyperparameters in the baseline progression}\item{param_Sigma_e}{1xI, the ground truth Gaussian noises.}\item{param_sigma2_f}{1xD, the ground truth hyperparameters for the Gaussian process treatment effect noises.}\item{param_rho_f}{1xD, the ground truth hyperparameters for the Gaussian process treatment effect noises.}
}
\description{
This functions generates heterogenous longitudinal data, in which baseline progressions are simulated from 3 clusters, two treatment types are generated (one type has positive effects, the other has negative effects), and the treatment responses for each type are respectively simulated from 3 clusters.
}
\examples{
simulate_data(I = 200, minT = 12, maxT = 24, unit = 60,
             minDur = 5, maxDur = 20, minADur = 60, maxADur = 80, W = c(60,70), 
              prior_unit = 1, trajectory_plot = TRUE)
}
\author{
Yanbo Xu
}
